+++ packages/lesmis-server/package.json
@@ -29,15 +29,14 @@
   },
   "dependencies": {
     "@tupaia/access-policy": "3.0.0",
+    "@tupaia/api-client": "3.1.0",
     "@tupaia/database": "1.0.0",
     "@tupaia/server-boilerplate": "1.0.0",
     "@tupaia/utils": "1.0.0",
     "camelcase-keys": "^6.2.2",
-    "cookies": "^0.8.0",
     "dotenv": "^8.2.0",
     "express": "^4.16.2",
     "lodash.uniqby": "^4.7.0",
-    "puppeteer": "^15.4.0",
     "winston": "^3.2.1"
   }
 }
