+++ packages/lesmis-server/src/routes/PDFExportRoute.ts
@@ -30,62 +30,10 @@ export class PDFExportRoute extends Route<PDFExportRequest> {
     this.type = 'download';
   }
 
-  private extractSessionCookie = (): SessionCookies => {
-    const cookies = new Cookies(this.req, this.res);
-    const sessionCookieName = 'sessionCookie';
-    const sessionCookieValue = cookies.get(sessionCookieName) || '';
-    return { sessionCookieName, sessionCookieValue };
-  };
-
-  private verifyBody = (body: any): Body => {
-    const { pdfPageUrl } = body;
-    if (!pdfPageUrl || typeof pdfPageUrl !== 'string') {
-      throw new Error(`'pdfPageUrl' should be provided in request body, got: ${pdfPageUrl}`);
-    }
-    const location = new URL(pdfPageUrl);
-    if (!location.hostname.endsWith('.tupaia.org') && !location.hostname.endsWith('localhost')) {
-      throw new Error(`'pdfPageUrl' is not valid, got: ${pdfPageUrl}`);
-    }
-    return { pdfPageUrl };
-  };
-
-  private exportPDF = async (): Promise<Buffer> => {
-    const { sessionCookieName, sessionCookieValue } = this.extractSessionCookie();
-    const { pdfPageUrl } = this.verifyBody(this.req.body);
-    const { host: apiDomain } = this.req.headers;
-    const location = new URL(pdfPageUrl);
-
-    let browser;
-    let result;
-
-    try {
-      browser = await puppeteer.launch();
-      const page = await browser.newPage();
-      await page.setCookie({
-        name: sessionCookieName,
-        domain: apiDomain,
-        url: location.origin,
-        httpOnly: true,
-        value: sessionCookieValue,
-      });
-      await page.goto(pdfPageUrl, { timeout: 60000, waitUntil: 'networkidle0' });
-      result = await page.pdf({
-        format: 'a4',
-        printBackground: true,
-      });
-    } catch (e) {
-      throw new Error(`puppeteer error: ${(e as Error).message}`);
-    } finally {
-      if (browser) {
-        await browser.close();
-      }
-    }
-
-    return result;
-  };
-
   public async buildResponse() {
-    const buffer = await this.exportPDF();
+    const { pdfPageUrl } = this.req.body;
+
+    const { data: buffer } = await this.apiClient.pdfExport.getPDF(pdfPageUrl);
     this.res.set({
       'Content-Type': 'application/pdf',
       'Content-Length': buffer.length,
